refactor(frontend): break the @/utils barrel import cycle
VideoBlock, Quiz, QuizInVideo and markdownParser imported pure format
helpers (formatSeconds/formatTimestamp/timeAgo/escapeHTML) through the
'@/utils' barrel, which itself imports the editor tools that import those
components back -> circular deps that block tree-shaking and risk init-order
bugs. Move the four pure helpers into a leaf module '@/utils/format' (no
component/tool imports), re-export them from '@/utils' for existing consumers,
and point the four cycle-prone files at the leaf.

Found via `fallow dead-code`. vitest 103/103 green.
