+++ frontend/src/components/VideoBlock.vue
@@ -157,7 +157,7 @@
 <script setup>
 import { ref, onMounted, computed, watch, onBeforeUnmount } from 'vue'
 import { Button, Dialog, Dropdown } from 'frappe-ui'
-import { formatSeconds, formatTimestamp } from '@/utils'
+import { formatSeconds, formatTimestamp } from '@/utils/format'
 import { useSettings } from '@/stores/settings'
 import Play from '@/components/Icons/Play.vue'
 import QuizInVideo from '@/components/Modals/QuizInVideo.vue'
