+++ frontend/src/utils/format.js
@@ -1,0 +1,35 @@
+import { useTimeAgo } from '@vueuse/core'
+
+export function timeAgo(date) {
+	return useTimeAgo(date).value
+}
+
+export const formatSeconds = (time) => {
+	const minutes = Math.floor(time / 60)
+	const seconds = Math.floor(time % 60)
+	return `${minutes}:${seconds < 10 ? '0' : ''}${seconds}`
+}
+
+export const escapeHTML = (text) => {
+	if (!text) return ''
+	let escape_html_mapping = {
+		'<': '&lt;',
+		'>': '&gt;',
+		'"': '&quot;',
+		"'": '&#39;',
+		'`': '&#x60;',
+	}
+
+	return String(text).replace(
+		/[&<>"'`=]/g,
+		(char) => escape_html_mapping[char] || char
+	)
+}
+
+export const formatTimestamp = (seconds) => {
+	const date = new Date(seconds * 1000)
+	const hours = String(date.getUTCHours()).padStart(2, '0')
+	const minutes = String(date.getUTCMinutes()).padStart(2, '0')
+	const secs = String(date.getUTCSeconds()).padStart(2, '0')
+	return hours > 0 ? `${hours}:${minutes}:${secs}` : `${minutes}:${secs}`
+}
